refactor from domain models to entities for storage
