+++ Federation/hosts/.gitignore
@@ -1,1 +1,1 @@
-Migrations
+Migrations/
