+++ Federation/src/Domain/Models/GrantType.cs
@@ -1,0 +1,8 @@
+﻿// Copyright (c) 2014-2022 Sarin Na Wangkanai, All Rights Reserved.Apache License, Version 2.0
+
+namespace Wangkanai.Federation.Models;
+
+public static class GrantType
+{
+	public const string AuthorizationCode = "authorization_code";
+}
