+++ Federation/src/EntityFramework/Entities/IdentityGroup.cs
@@ -2,35 +2,33 @@
 
 using Wangkanai.Domain;
 
-namespace Wangkanai.Federation.Models;
+namespace Wangkanai.Federation.Entities;
 
-public class IdentityScope : IdentityScope<Guid>
+public class IdentityGroup : IdentityGroup<Guid>
 {
-	public IdentityScope()
+	public IdentityGroup()
 	{
 		Id = Guid.NewGuid();
 	}
 
-	public IdentityScope(string name) : this()
+	public IdentityGroup(string name) : this()
 	{
 		Name = name;
 	}
 }
 
-public class IdentityScope<TKey> : IAuditable
+public class IdentityGroup<TKey> : IAuditable
 	where TKey : IEquatable<TKey>
 {
-	public IdentityScope() { }
+	public IdentityGroup() { }
 
-	public IdentityScope(string name) : this()
+	public IdentityGroup(string name) : this()
 	{
 		Name = name;
 	}
 
 	public virtual TKey      Id               { get; set; } = default!;
 	public virtual string    Name             { get; set; }
-	public virtual string    DisplayName      { get; set; }
-	public virtual string    Description      { get; set; }
 	public virtual DateTime  Created          { get; set; } = DateTime.UtcNow;
 	public virtual DateTime? Updated          { get; set; }
 	public virtual string?   ConcurrencyStamp { get; set; } = Guid.NewGuid().ToString();
