+++ Federation/src/EntityFramework/Entities/IdentityPermission.cs
@@ -1,6 +1,6 @@
 // Copyright (c) 2014-2022 Sarin Na Wangkanai, All Rights Reserved.Apache License, Version 2.0
 
-namespace Wangkanai.Federation.Models;
+namespace Wangkanai.Federation.Entities;
 
 [Flags]
 public enum IdentityPermission
