+++ Federation/src/EntityFramework/Entities/IdentityResource.cs
@@ -2,33 +2,35 @@
 
 using Wangkanai.Domain;
 
-namespace Wangkanai.Federation.Models;
+namespace Wangkanai.Federation.Entities;
 
-public class IdentityGroup : IdentityGroup<Guid>
+public class IdentityResource : IdentityResource<Guid>
 {
-	public IdentityGroup()
+	public IdentityResource()
 	{
 		Id = Guid.NewGuid();
 	}
 
-	public IdentityGroup(string name) : this()
+	public IdentityResource(string name) : this()
 	{
 		Name = name;
 	}
 }
 
-public class IdentityGroup<TKey> : IAuditable
+public class IdentityResource<TKey> : IAuditable
 	where TKey : IEquatable<TKey>
 {
-	public IdentityGroup() { }
+	public IdentityResource() { }
 
-	public IdentityGroup(string name) : this()
+	public IdentityResource(string name) : this()
 	{
 		Name = name;
 	}
 
 	public virtual TKey      Id               { get; set; } = default!;
 	public virtual string    Name             { get; set; }
+	public virtual string    DisplayName      { get; set; }
+	public virtual string    Description      { get; set; }
 	public virtual DateTime  Created          { get; set; } = DateTime.UtcNow;
 	public virtual DateTime? Updated          { get; set; }
 	public virtual string?   ConcurrencyStamp { get; set; } = Guid.NewGuid().ToString();
